renaming/refactoring for doc gen purposes
