+++ warp10/src/main/java/io/warp10/script/functions/shape/HULLSHAPE.java
@@ -14,7 +14,7 @@
 //   limitations under the License.
 //
 
-package io.warp10.script.functions;
+package io.warp10.script.functions.shape;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -24,75 +24,68 @@ import io.warp10.script.WarpScriptException;
 import io.warp10.script.WarpScriptStack;
 import io.warp10.script.formatted.FormattedWarpScriptFunction;
 
-public class CHECKSHAPE extends FormattedWarpScriptFunction {
+public class HULLSHAPE extends FormattedWarpScriptFunction {
 
   public static final String LIST = "list";
-  public static final String RES = "result";
+  public static final String SHAPE = "shape";
 
   private final Arguments args;
   private final Arguments output;
   protected Arguments getArguments() { return args; }
   protected Arguments getOutput() { return output; }
 
-  public CHECKSHAPE(String name) {
+  public HULLSHAPE(String name) {
     super(name);
 
-    getDocstring().append("Return a BOOLEAN indicating whether an input list and its nested lists sizes are coherent together to form a tensor (or multidimensional array).");
+    getDocstring().append("Return the shape of a tensor (or multidimensional array) that would be able to contain all the values of an input nested list. The size of the returned shape is equal to the deepest level of nesting plus one. Its i-th value is equal to the size of the largest list that is nested i levels deep.");
 
     args = new ArgumentsBuilder()
       .addArgument(List.class, LIST, "The input list.")
       .build();
 
     output = new ArgumentsBuilder()
-      .addListArgument(Long.class, RES, "true or false.")
+      .addListArgument(Long.class, SHAPE, "The hull shape of the input list.")
       .build();
   }
 
   @Override
   protected WarpScriptStack apply(Map<String, Object> formattedArgs, WarpScriptStack stack) throws WarpScriptException {
     List list = (List) formattedArgs.get(LIST);
-    List<Long> candidateShape = SHAPE.candidate_shape(list);
-    stack.push(recValidateShape(list, candidateShape));
+    stack.push(recHullShape(list));
     return stack;
   }
 
-  private static Boolean hasNestedList(List list) {
+  private List<Long> maximizeHull(List<Long> a, List<Long> b) {
+    List<Long> smaller = a.size() < b.size() ? a : b;
+    List<Long> taller = a.size() < b.size() ? b : a;
+
+    if (0 == smaller.size()) {
+      return taller;
+    }
+
+    List<Long> res = new ArrayList<Long>(taller.size());
+    while (smaller.size() > 0) {
+      Long x = smaller.remove(0);
+      Long y = taller.remove(0);
+      res.add(x > y ? x : y);
+    }
+
+    res.addAll(taller);
+    return res;
+  }
+
+  private List<Long> recHullShape(List list) {
+    List<Long> res = new ArrayList<Long>();
+    res.add((long) list.size());
+
+    List<Long> hull = new ArrayList<>();
     for (Object el: list) {
       if (el instanceof List) {
-        return true;
+        hull = maximizeHull(hull, recHullShape((List) el));
       }
     }
 
-    return false;
+    res.addAll(hull);
+    return res;
   }
-
-  static Boolean recValidateShape(List list, List<Long> candidateShape) {
-    List<Long> copyShape =  new ArrayList<Long>(candidateShape);
-
-    if (list.size() != copyShape.remove(0)) {
-      return false;
-    }
-
-    if (!hasNestedList(list)) {
-      return 0 == copyShape.size();
-
-    } else if (0 == copyShape.size()) {
-      return false;
-
-    } else {
-
-      for (Object el: list) {
-        if (!(el instanceof List)) {
-          return false;
-        }
-
-        if (!recValidateShape((List) el, copyShape)) {
-          return false;
-        }
-      }
-
-      return true;
-    }
-  }
-
 }
